Added Environment class/basic functionality

+++ BCEngine/src/BCSim/Physics.cpp
@@ -2,15 +2,15 @@
 
 namespace BCSim
 {
-	Vector2f::Vector2f()
+	Vector2d::Vector2d()
 		:x(0.0f), y(0.0f)
 	{}
 
-	Vector2f::Vector2f(float x, float y)
+	Vector2d::Vector2d(float x, float y)
 		:x(x), y(y)
 	{}
 
-	void Vector2f::PrintData()
+	void Vector2d::PrintData()
 	{
 		std::cout << x << ", " << y << std::endl;
 	}
