+++ BCEngine/src/BCSim/Physics.h
@@ -6,11 +6,11 @@ namespace BCSim
 {
 	//TODO: Rename this (Maybe Double_Vector? will get confusing with actual 2D vector)
 	//Or rework to use doubles, going back to Vector2f for now
-	class Vector2f
+	class Vector2d
 	{
 	public:
-		Vector2f();
-		Vector2f(float x, float y);
+		Vector2d();
+		Vector2d(float x, float y);
 		void PrintData();
 		float x, y;
 	};
